neler oldu neler bitti.. Paylasim.java isimli dosyamiza baska uygulamalardan gelen paylasma hecetinin bilgileri analiz edilip degerlendiriliyor. pardon analiz edilip degerlendirilmiyor.sadece o paylasilan objenin uri si icerigi,text se yazisi falan,ha bu ne isimize yaricak,bu sayede kisayollar daha net bi sekilde widget e aktarilabilinecek,paylasilan objenin kisayolu yapilacak haydi kal saglicakla..

+++ app/src/main/java/com/natrollus/nobet/Widget.java
@@ -73,7 +73,7 @@ public class Widget extends AppWidgetProvider {
 	}
 
 	private void sayac(int artim) {
-		int kac = degerGetir(context,"kac",0);
+		int kac = (int) degerGetir(context,"kac",0);
 		degerEkle(context,"kac",(int) degerGetir(context,"kac",0)+artim);
 		rv.setTextViewText(R.id.sayac, "" + kac);
 	}
